// src/config/payload-types.ts
export interface User {
  id: number
  name?: string | null
  updatedAt: string
  createdAt: string
  email: string
  resetPasswordToken?: string | null
  resetPasswordExpiration?: string | null
  salt?: string | null
  hash?: string | null
  loginAttempts?: number | null
  lockUntil?: string | null
  password?: string | null
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "categories".
 */
export interface Category {
  id: number
  title: string
  parent?: (number | null) | Category
  breadcrumbs?:
    | {
        doc?: (number | null) | Category
        url?: string | null
        label?: string | null
        id?: string | null
      }[]
    | null
  updatedAt: string
  createdAt: string
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "posts".
 */
export interface Post {
  id: number
  title: string
  lang: string
  content: {
    root: {
      type: string
      children: {
        type: string
        version: number
        [k: string]: unknown
      }[]
      direction: ("ltr" | "rtl") | null
      format: "left" | "start" | "center" | "right" | "end" | "justify" | ""
      indent: number
      version: number
    }
    [k: string]: unknown
  }
  relatedPosts?: (number | Post)[] | null
  categories?: (number | Category)[] | null
  meta?: {
    title?: string | null
    image?: (number | null) | Media
    description?: string | null
  }
  publishedAt?: string | null
  authors?: (number | User)[] | null
  populatedAuthors?:
    | {
        id?: string | null
        name?: string | null
      }[]
    | null
  slug?: string | null
  slugLock?: boolean | null
  updatedAt: string
  createdAt: string
  _status?: ("draft" | "published") | null
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "media".
 */
export interface Media {
  id: number
  alt?: string | null
  caption?: {
    root: {
      type: string
      children: {
        type: string
        version: number
        [k: string]: unknown
      }[]
      direction: ("ltr" | "rtl") | null
      format: "left" | "start" | "center" | "right" | "end" | "justify" | ""
      indent: number
      version: number
    }
    [k: string]: unknown
  } | null
  updatedAt: string
  createdAt: string
  url?: string | null
  thumbnailURL?: string | null
  filename?: string | null
  mimeType?: string | null
  filesize?: number | null
  width?: number | null
  height?: number | null
  focalX?: number | null
  focalY?: number | null
  sizes?: {
    thumbnail?: {
      url?: string | null
      width?: number | null
      height?: number | null
      mimeType?: string | null
      filesize?: number | null
      filename?: string | null
    }
    square?: {
      url?: string | null
      width?: number | null
      height?: number | null
      mimeType?: string | null
      filesize?: number | null
      filename?: string | null
    }
    small?: {
      url?: string | null
      width?: number | null
      height?: number | null
      mimeType?: string | null
      filesize?: number | null
      filename?: string | null
    }
    medium?: {
      url?: string | null
      width?: number | null
      height?: number | null
      mimeType?: string | null
      filesize?: number | null
      filename?: string | null
    }
    large?: {
      url?: string | null
      width?: number | null
      height?: number | null
      mimeType?: string | null
      filesize?: number | null
      filename?: string | null
    }
    xlarge?: {
      url?: string | null
      width?: number | null
      height?: number | null
      mimeType?: string | null
      filesize?: number | null
      filename?: string | null
    }
  }
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "pages".
 */
export interface Page {
  id: number
  title: string
  lang: string
  hero: {
    type: "none" | "highImpact" | "mediumImpact" | "lowImpact"
    richText?: {
      root: {
        type: string
        children: {
          type: string
          version: number
          [k: string]: unknown
        }[]
        direction: ("ltr" | "rtl") | null
        format: "left" | "start" | "center" | "right" | "end" | "justify" | ""
        indent: number
        version: number
      }
      [k: string]: unknown
    } | null
    links?:
      | {
          link: {
            type?: ("reference" | "custom") | null
            newTab?: boolean | null
            reference?: {
              relationTo: "pages"
              value: number | Page
            } | null
            url?: string | null
            label: string
            appearance?: ("default" | "outline") | null
          }
          id?: string | null
        }[]
      | null
    media?: (number | null) | Media
  }
  layout: (
    | CallToActionBlock
    | ContentBlock
    | MediaBlock
    | ArchiveBlock
    | FormBlock
  )[]
  meta?: {
    title?: string | null
    image?: (number | null) | Media
    description?: string | null
  }
  publishedAt?: string | null
  slug?: string | null
  slugLock?: boolean | null
  updatedAt: string
  createdAt: string
  _status?: ("draft" | "published") | null
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "CallToActionBlock".
 */
export interface CallToActionBlock {
  richText?: {
    root: {
      type: string
      children: {
        type: string
        version: number
        [k: string]: unknown
      }[]
      direction: ("ltr" | "rtl") | null
      format: "left" | "start" | "center" | "right" | "end" | "justify" | ""
      indent: number
      version: number
    }
    [k: string]: unknown
  } | null
  links?:
    | {
        link: {
          type?: ("reference" | "custom") | null
          newTab?: boolean | null
          reference?: {
            relationTo: "pages"
            value: number | Page
          } | null
          url?: string | null
          label: string
          appearance?: ("default" | "outline") | null
        }
        id?: string | null
      }[]
    | null
  id?: string | null
  blockName?: string | null
  blockType: "cta"
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "ContentBlock".
 */
export interface ContentBlock {
  columns?:
    | {
        size?: ("oneThird" | "half" | "twoThirds" | "full") | null
        richText?: {
          root: {
            type: string
            children: {
              type: string
              version: number
              [k: string]: unknown
            }[]
            direction: ("ltr" | "rtl") | null
            format:
              | "left"
              | "start"
              | "center"
              | "right"
              | "end"
              | "justify"
              | ""
            indent: number
            version: number
          }
          [k: string]: unknown
        } | null
        enableLink?: boolean | null
        link?: {
          type?: ("reference" | "custom") | null
          newTab?: boolean | null
          reference?: {
            relationTo: "pages"
            value: number | Page
          } | null
          url?: string | null
          label: string
          appearance?: ("default" | "outline") | null
        }
        id?: string | null
      }[]
    | null
  id?: string | null
  blockName?: string | null
  blockType: "content"
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "MediaBlock".
 */
export interface MediaBlock {
  media: number | Media
  id?: string | null
  blockName?: string | null
  blockType: "mediaBlock"
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "ArchiveBlock".
 */
export interface ArchiveBlock {
  introContent?: {
    root: {
      type: string
      children: {
        type: string
        version: number
        [k: string]: unknown
      }[]
      direction: ("ltr" | "rtl") | null
      format: "left" | "start" | "center" | "right" | "end" | "justify" | ""
      indent: number
      version: number
    }
    [k: string]: unknown
  } | null
  populateBy?: ("collection" | "selection") | null
  relationTo?: "posts" | null
  categories?: (number | Category)[] | null
  limit?: number | null
  selectedDocs?:
    | {
        relationTo: "posts"
        value: number | Post
      }[]
    | null
  id?: string | null
  blockName?: string | null
  blockType: "archive"
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "FormBlock".
 */
export interface FormBlock {
  form: number | Form
  enableIntro?: boolean | null
  introContent?: {
    root: {
      type: string
      children: {
        type: string
        version: number
        [k: string]: unknown
      }[]
      direction: ("ltr" | "rtl") | null
      format: "left" | "start" | "center" | "right" | "end" | "justify" | ""
      indent: number
      version: number
    }
    [k: string]: unknown
  } | null
  id?: string | null
  blockName?: string | null
  blockType: "formBlock"
}